Update Admin 4/10, except Doanh thu

+++ Controllers/Manage/ManagePhonesController.cs
@@ -7,6 +7,8 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using old_phone.Models;
+using old_phone.Common;
+using PagedList;
 
 namespace old_phone.Controllers.Manage
 {
@@ -15,13 +17,34 @@ namespace old_phone.Controllers.Manage
         private OldPhoneEntities db = new OldPhoneEntities();
 
         // GET: ManagePhones
-        public ActionResult Index()
+        [AuthorizeCheck(RequiredRole = 2)]
+        public ActionResult Index(string searchQuery, int? companyId, int? page)
         {
-            var phones = db.Phones.Include(p => p.Product);
-            return View(phones.ToList());
+            var phones = db.Phones
+                .Include(p => p.Product)
+                .Include(p => p.Product.Company)
+                .OrderByDescending(p => p.product_id)
+                .AsQueryable();
+            if (!string.IsNullOrEmpty(searchQuery))
+            {
+                phones = phones.Where(p => p.Product.product_name.Contains(searchQuery));
+            }
+            if (companyId.HasValue)
+            {
+                phones = phones.Where(p => p.Product.company_id == companyId.Value);
+            }
+            int pageSize = 20;
+            int pageNumber = (page ?? 1);
+            ViewBag.CurrentFilter = searchQuery;
+            ViewBag.CurrentCompanyID = companyId;
+            ViewBag.CompanyList = new SelectList(db.Companies, "company_id", "company_name",companyId);
+
+            var pagedPhones = phones.ToPagedList(pageNumber, pageSize);
+            return View(pagedPhones);
         }
 
         // GET: ManagePhones/Details/5
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -37,6 +60,7 @@ namespace old_phone.Controllers.Manage
         }
 
         // GET: ManagePhones/Create
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult Create()
         {
             ViewBag.product_id = new SelectList(db.Products, "product_id", "product_name");
@@ -48,6 +72,7 @@ namespace old_phone.Controllers.Manage
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult Create([Bind(Include = "product_id,phone_chip,phone_screen_size,phone_front_cam,phone_behind_cam,phone_battery,phone_system,phone_charging_port,phone_sim_card,phone_nfc,phone_ear_phone,phone_memory_card")] Phone phone)
         {
             if (ModelState.IsValid)
@@ -62,6 +87,7 @@ namespace old_phone.Controllers.Manage
         }
 
         // GET: ManagePhones/Edit/5
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -82,6 +108,7 @@ namespace old_phone.Controllers.Manage
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult Edit([Bind(Include = "product_id,phone_chip,phone_screen_size,phone_front_cam,phone_behind_cam,phone_battery,phone_system,phone_charging_port,phone_sim_card,phone_nfc,phone_ear_phone,phone_memory_card")] Phone phone)
         {
             if (ModelState.IsValid)
@@ -95,6 +122,7 @@ namespace old_phone.Controllers.Manage
         }
 
         // GET: ManagePhones/Delete/5
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -112,12 +140,34 @@ namespace old_phone.Controllers.Manage
         // POST: ManagePhones/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [AuthorizeCheck(RequiredRole = 2)]
         public ActionResult DeleteConfirmed(int id)
         {
             Phone phone = db.Phones.Find(id);
-            db.Phones.Remove(phone);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            var list_variant = db.Variant_Phone.Where(v => v.product_id == id).ToList();
+            try
+            {
+                if (list_variant.Count() == 0)
+                {
+                    db.Phones.Remove(phone);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return View("~/Views/Shared/Error_Delete.cshtml");
+                }
+            }
+            catch (System.Data.Entity.Infrastructure.DbUpdateException)
+            {
+                return View("~/Views/Shared/Error_Delete.cshtml");
+            }
+            catch (Exception ex) 
+            {
+                ViewBag.ErrorMessage= ex.Message;
+                return View("~/Views/Shared/Error_Delete.cshtml");
+            }
+
         }
 
         protected override void Dispose(bool disposing)
